requirements.txt
pymongo==4.6.1
python-telegram-bot==20.7